+++ tests/test_event_processor.py
@@ -41,7 +41,19 @@ MEMBER_LEVEL_4 = MemberLevel(Id=444, Url="http://example.com/444")
 SAMPLE_RESTRICTION = EventRestriction(
     name="Members Only",
     pattern=re.compile("members[ -]*only", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
     member_levels=[MEMBER_LEVEL_2, MEMBER_LEVEL_3, MEMBER_LEVEL_4],
+    guest_policy="Disabled",
+)
+
+DEFAULT_RESTRICTION = EventRestriction(
+    name="RSVP",
+    pattern=re.compile("^", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
+    member_levels=[],
+    guest_policy="Disabled",
 )
 
 SAMPLE_MEMBER_LEVELS = [
@@ -145,7 +157,7 @@ def mock_apricot_api(mocker):
 @pytest.fixture()
 def event_registration_type_maker():
     """Return an event registration type maker."""
-    return EventRegistrationTypeMaker([SAMPLE_RESTRICTION])
+    return EventRegistrationTypeMaker([SAMPLE_RESTRICTION, DEFAULT_RESTRICTION])
 
 
 @pytest.fixture()