+++ tests/test_event_registration_type.py
@@ -19,13 +19,46 @@ SAMPLE_MEMBER_LEVELS_JSON = [
 SAMPLE_ALL_LEVELS_RESTRICTION = EventRestriction(
     name="Members Only",
     pattern=re.compile("members[ -]*only", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
     member_levels=ALL_LEVELS,
+    guest_policy="CollectContactDetails",
 )
 
 SAMPLE_NAMED_LEVELS_RESTRICTION = EventRestriction(
     name="Key Members Only",
     pattern=re.compile("key +members +only", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
     member_levels=[MEMBER_LEVEL_2],
+    guest_policy="Disabled",
+)
+
+SAMPLE_FREE_RESTRICTION = EventRestriction(
+    name="Free",
+    pattern=re.compile("sample", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=False,
+    member_levels=ALL_LEVELS,
+    guest_policy="Disabled",
+)
+
+SAMPLE_PAID_RESTRICTION = EventRestriction(
+    name="Paid",
+    pattern=re.compile("sample", re.IGNORECASE),
+    match_free_events=False,
+    match_paid_events=True,
+    member_levels=ALL_LEVELS,
+    guest_policy="Disabled",
+)
+
+DEFAULT_RESTRICTION = EventRestriction(
+    name="RSVP",
+    pattern=re.compile("^", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
+    member_levels=[],
+    guest_policy="Disabled",
 )
 
 EXPECTED_APRICOT_JSON = {
@@ -69,7 +102,7 @@ EXPECTED_MEMBERS_ONLY_JSON = {
     "Description": "",
     "EventId": 12345,
     "GuestPrice": 78.9,
-    "GuestRegistrationPolicy": "Disabled",
+    "GuestRegistrationPolicy": "CollectContactDetails",
     "IsEnabled": True,
     "IsWaitlistEnabled": False,
     "MaximumRegistrantsCount": 6,
@@ -81,7 +114,13 @@ EXPECTED_MEMBERS_ONLY_JSON = {
 @pytest.fixture()
 def event_registration_type_maker():
     """Return an event registration type maker."""
-    restrictions = [SAMPLE_NAMED_LEVELS_RESTRICTION, SAMPLE_ALL_LEVELS_RESTRICTION]
+    restrictions = [
+        SAMPLE_NAMED_LEVELS_RESTRICTION,
+        SAMPLE_ALL_LEVELS_RESTRICTION,
+        SAMPLE_FREE_RESTRICTION,
+        SAMPLE_PAID_RESTRICTION,
+        DEFAULT_RESTRICTION,
+    ]
     return EventRegistrationTypeMaker(restrictions)
 
 
@@ -91,7 +130,7 @@ def test_make_unrestricted_apricot_type_for_json(
     """Test converting an unrestricted Wild Apricot RSVP event registration
     type into dictionaries and lists suitable for conversion to JSON."""
     reg_type = event_registration_type_maker.make_unrestricted_apricot_type(
-        12345, 6, 78.9
+        12345, 6, 78.9, DEFAULT_RESTRICTION
     )
     assert reg_type.for_json() == EXPECTED_APRICOT_JSON
 
@@ -112,18 +151,36 @@ def test_make_restricted_apricot_type_for_json(event_registration_type_maker):
     assert reg_type.for_json() == EXPECTED_MEMBERS_ONLY_JSON
 
 
-def test_choose_event_restriction_no_match(event_registration_type_maker):
-    """Test choosing an event restriction when no pattern matches the event title."""
+def test_choose_event_restriction_default_match(event_registration_type_maker):
+    """Test choosing an event restriction when the default restriction matches."""
     event_title = "Mending Monday"
-    restriction = event_registration_type_maker.choose_event_restriction(event_title)
-    assert restriction is None
+    restriction = event_registration_type_maker.choose_event_restriction(event_title, 0)
+    assert restriction is DEFAULT_RESTRICTION
 
 
 def test_choose_event_restriction_first_match(event_registration_type_maker):
     """Test choosing an event restriction when no pattern matches the event title."""
     event_title = "Mending Monday (Members Only)"
-    restriction = event_registration_type_maker.choose_event_restriction(event_title)
+    restriction = event_registration_type_maker.choose_event_restriction(
+        event_title, 35.0
+    )
     assert restriction == SAMPLE_ALL_LEVELS_RESTRICTION
+
+
+def test_choose_event_restriction_free_match(event_registration_type_maker):
+    """Test choosing an event restriction when the title matches and the event is free."""
+    event_title = "Sample"
+    restriction = event_registration_type_maker.choose_event_restriction(event_title, 0)
+    assert restriction == SAMPLE_FREE_RESTRICTION
+
+
+def test_choose_event_restriction_paid_match(event_registration_type_maker):
+    """Test choosing an event restriction when the title matches and the event is paid."""
+    event_title = "Sample"
+    restriction = event_registration_type_maker.choose_event_restriction(
+        event_title, 35.0
+    )
+    assert restriction == SAMPLE_PAID_RESTRICTION
 
 
 def test_make_apricot_type_unrestricted(event_registration_type_maker):