+++ tests/test_event_restriction_loader.py
@@ -4,7 +4,11 @@ from meetup2apricot.event_restriction_loader import (
     EventRestrictionLoader,
     EventRestriction,
 )
-from meetup2apricot.exceptions import InvalidRestrictionPattern
+from meetup2apricot.exceptions import (
+    InvalidGuestPolicyError,
+    InvalidPriceRestriction,
+    InvalidRestrictionPattern,
+)
 from meetup2apricot.member_level_manager import MemberLevel, MemberLevelManager
 import re
 import pytest
@@ -14,29 +18,45 @@ MEMBER_LEVEL_1 = MemberLevel(Id=111, Url="http://example.com/111")
 MEMBER_LEVEL_2 = MemberLevel(Id=222, Url="http://example.com/222")
 MEMBER_LEVEL_3 = MemberLevel(Id=333, Url="http://example.com/333")
 
-ALL_LEVELS = [MEMBER_LEVEL_1, MEMBER_LEVEL_2, MEMBER_LEVEL_3]
-
 SAMPLE_ALL_LEVELS_RESTRICTION_JSON = {
     "name": "Members Only",
     "pattern": "members[ -]*only",
+    "price": "free",
+    "levels": ["Key", "Associate", "Family"],
 }
 
 SAMPLE_NAMED_LEVELS_RESTRICTION_JSON = {
     "name": "Key Members Only",
     "pattern": "key +members +only",
+    "price": "paid",
     "levels": ["Key", "Family"],
 }
 
 SAMPLE_ALL_LEVELS_RESTRICTION = EventRestriction(
     name="Members Only",
     pattern=re.compile("members[ -]*only", re.IGNORECASE),
-    member_levels=ALL_LEVELS,
+    match_free_events=True,
+    match_paid_events=False,
+    member_levels=[MEMBER_LEVEL_1, MEMBER_LEVEL_2, MEMBER_LEVEL_3],
+    guest_policy="Disabled",
 )
 
 SAMPLE_NAMED_LEVELS_RESTRICTION = EventRestriction(
     name="Key Members Only",
     pattern=re.compile("key +members +only", re.IGNORECASE),
+    match_free_events=False,
+    match_paid_events=True,
     member_levels=[MEMBER_LEVEL_1, MEMBER_LEVEL_3],
+    guest_policy="Disabled",
+)
+
+EXPECTED_DEFAULT_RESTRICTION = EventRestriction(
+    name="RSVP",
+    pattern=re.compile("^", re.IGNORECASE),
+    match_free_events=True,
+    match_paid_events=True,
+    member_levels=[],
+    guest_policy="Disabled",
 )
 
 
@@ -48,7 +68,7 @@ def member_level_manager():
         "Associate": MEMBER_LEVEL_2,
         "Family": MEMBER_LEVEL_3,
     }
-    return MemberLevelManager(ALL_LEVELS, named_levels)
+    return MemberLevelManager(named_levels)
 
 
 @pytest.fixture()
@@ -71,6 +91,42 @@ def test_compile_pattern_invalid():
         pattern = EventRestrictionLoader.compile_pattern("[")
 
 
+def test_parse_price_invalid():
+    """Test parsing an invalid price restriction."""
+    expected_message_pattern = (
+        r'Event price restriction "oops" must be "free", "paid", or omitted'
+    )
+    with pytest.raises(InvalidPriceRestriction, match=expected_message_pattern):
+        EventRestrictionLoader.parse_price("oops")
+
+
+def test_parse_guests_policy_invalid():
+    """Test parsing an invalid guest policy."""
+    expected_message_pattern = (
+        r'Guest policy "oops" must be "count", "contact", "full", or omitted'
+    )
+    with pytest.raises(InvalidGuestPolicyError, match=expected_message_pattern):
+        EventRestrictionLoader.parse_guest_policy("oops")
+
+
+def test_parse_guests_policy_count():
+    """Test parsing a "count" guest policy."""
+    guest_policy = EventRestrictionLoader.parse_guest_policy("count")
+    assert guest_policy == "NumberOfGuests"
+
+
+def test_parse_guests_policy_contact():
+    """Test parsing a "contact" guest policy."""
+    guest_policy = EventRestrictionLoader.parse_guest_policy("contact")
+    assert guest_policy == "CollectContactDetails"
+
+
+def test_parse_guests_policy_full():
+    """Test parsing a "full" guest policy."""
+    guest_policy = EventRestrictionLoader.parse_guest_policy("full")
+    assert guest_policy == "CollectFullInfo"
+
+
 @pytest.mark.parametrize(
     "raw_level_list, expected_level_list",
     [
@@ -90,7 +146,7 @@ def test_clean_level_names(raw_level_list, expected_level_list):
 
 def test_lookup_member_levels_empty(event_restriction_loader):
     """Test looking up an empty list of member level names."""
-    assert event_restriction_loader.lookup_member_levels([]) == ALL_LEVELS
+    assert event_restriction_loader.lookup_member_levels([]) == []
 
 
 def test_lookup_member_levels_list(event_restriction_loader):
@@ -109,11 +165,17 @@ def test_load_restriction_all(event_restriction_loader):
 
 
 def test_load_restriction_named(event_restriction_loader):
-    """Test loading a restriction that namedows named member levels."""
+    """Test loading a restriction with named member levels."""
     restriction = event_restriction_loader.load_restriction(
         SAMPLE_NAMED_LEVELS_RESTRICTION_JSON
     )
     assert restriction == SAMPLE_NAMED_LEVELS_RESTRICTION
+
+
+def test_load_restriction_default(event_restriction_loader):
+    """Test default values from loading a restriction with no details."""
+    restriction = event_restriction_loader.load_restriction({})
+    assert restriction == EXPECTED_DEFAULT_RESTRICTION
 
 
 def test_load(event_restriction_loader):
@@ -124,6 +186,7 @@ def test_load(event_restriction_loader):
     assert restrictions == [
         SAMPLE_NAMED_LEVELS_RESTRICTION,
         SAMPLE_ALL_LEVELS_RESTRICTION,
+        EXPECTED_DEFAULT_RESTRICTION,
     ]