+++ HISTORY.rst
@@ -12,8 +12,11 @@ Future History
 Recent History
 ~~~~~~~~~~~~~~
 
-.. Next Release
-.. ------------------
+Next Release
+------------------
+
+* Add registration restrictions based on price.
+* Make guest registration policy configurable.
 
 1.8.1 (2021-04-09)
 ------------------